+++ packages/vidstack/src/components/ui/buttons/caption-button.ts
@@ -58,8 +58,8 @@ export class CaptionButton extends Component<CaptionButtonProps> {
   }
 
   private _isHidden() {
-    const { textTracks } = this._media.$state;
-    return textTracks().filter(isTrackCaptionKind).length == 0;
+    const { hasCaptions } = this._media.$state;
+    return !hasCaptions();
   }
 
   private _getLabel() {